Refactor HomeClient and useHomepageData for improved data handling
- Simplified data age calculation in HomeClient by using useMemo for better performance.
- Extracted topic computation logic into a separate utility function, computeTopics, to enhance code organization and readability.
- Updated useHomepageData to include initialDataUpdatedAt for better cache management and removed unnecessary refetchOnMount logic.
- Cleaned up unused hooks and code related to homepage data age, streamlining the overall implementation.
